+++ packages/myst-ext-button/README.md
@@ -1,0 +1,3 @@
+# myst-ext-button
+
+`mystmd` extension for `button` role
